docs(pipes): update structure and move to helpers
