--- projects/canopy/src/lib/pipes/camel-case/camel-case.notes.ts
@@ -1,13 +1,0 @@
-export const notes = `
-This pipe allows for a string to be transformed into camel case.
-
-## Usage
-Import the pipe in your module:
-
-~~~
-@NgModule({
-  ...
-  imports: [ ..., LgCamelCasePipeModule ],
-})
-~~~
-`;
